pyproject.toml
[project]
name = "helmet-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "av>=12.3.0",
    "gdown>=5.2.2",
    "opencv-python>=4.13.0.92",
    "pillow>=10.4.0",
    "streamlit>=1.40.1",
    "streamlit-webrtc>=0.48.1",
    "ultralytics>=8.4.41",
]
